abstract spawn manager and target vehicle manager
